+++ Dockerfile
@@ -1,4 +1,4 @@
-ARG BASE_IMAGE=pennlinc/qsiprep-base:20260415
+ARG BASE_IMAGE=pennlinc/qsiprep-base:20260804
 ARG DWIDENOISE2_COMMIT=892d5e8dd8f453ce1a561878f7dcb3998ae258ba
 ARG MRTRIX3_DWIDENOISE2_COMMIT=fa6ee952913fbc1df79aeca745600852155f533f
